+++ packages/scoped-custom-element-registry/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- ## Unreleased -->
 
+- Run `connectedCallback` on upgraded elements. Fixes #442.
 - Checks if ShadowRoot prototype supports the createElement method to determine if the polyfill should be applied or not.
 
 ## [0.0.1] - 2021-02-18